Support dumping game class files into a Jar with everything public.
We now support generating a jar file that contains all PZ class files
modified to be public and non-final. This also applies to fields and
methods of those classes.

Additionally, at runtime, the same transformation will be applied before
classes are loaded. This is intended to facilitate an easier means of
working with game classes since they can be subclassed and
data/functions can be directly accessed.