Support 3 new arguments: --server, --main and --mods
For the benefit of commandline use, we now support 3 new options to make
it easier to launch PZHook with preconfigured options.

--server will result in Project Zomboid's zomboid.network.GameServer
main being called instead of the usual client main for people playing
the game.

--main <classpath> will result in whatever <classpath> is being invoked
as main instead. Primarily aimed at people who need to do their own
setup before allowing the game to execute. Previously this was
achieveable by hooking the game's main, but this is of course "nicer".

--mods <mod>;<mod> will result in PZHook enabling all mods whose name
matches whatever is passed in this semicolon delimited list. It is
mutually exclusive with the --i-trust-all-mods option.